Change the data_model structure
The new structure is:

module_name
├── dm.py
├── __init__.py
└── strategy.py

where __init__.py imports and exposes data_model and tactics from the
rest of the module
