--- src/fuddly/data_models/__init__.py
@@ -1,1 +1,0 @@
-
