--- src/fuddly/data_models/file_formats/__init__.py
@@ -1,1 +1,0 @@
-
